[feat/#49] ArchiveView 구현
