+++ Solply/Solply/Presentation/Archive/ArchiveView/ArchiveView.swift
@@ -1,0 +1,33 @@
+//
+//  ArchiveView.swift
+//  Solply
+//
+//  Created by 이수용 on 7/9/25.
+//
+
+import SwiftUI
+
+struct ArchiveView: View {
+    
+    @EnvironmentObject var appCoordinator: AppCoordinator
+    @StateObject var store = ArchiveStore()
+    
+    var body: some View {
+        
+        VStack(alignment: .leading, spacing: 0) {
+            ArchiveBar { category in
+                store.dispatch(.toggleArchiveBar(archiveCategory: category))
+            }
+            
+            ArchiveEmptyView(archiveCategory: store.state.selectedCategory)
+        }
+        .frame(maxHeight: .infinity, alignment: .top)
+        .customNavigationBar(.archive(backAction: appCoordinator.goBack))
+            
+    }
+}
+
+#Preview {
+    ArchiveView()
+        .environmentObject(AppCoordinator())
+}
